+++ examples/hello/server.js
@@ -3,7 +3,7 @@ import React from 'react';
 import ReactDOMServer from 'react-dom/server';
 import express from 'express';
 
-import HelloWorldEmail from './HelloWorldEmail.jsx';
+import {HelloWorldEmail} from './HelloWorldEmail.jsx';
 
 
 const server = express();